bug: 【流水线】JOB日志缺失，如启动镜像相关的日志 #3493
